requirements.txt
django
djangorestframework
django-cors-headers
drf-yasg
coreapi
django-rest-passwordreset
Pillow